Create constant wall thickness radius for opponent card
Constant wall thickness is an important aesthetic feature in my opinion as a mechanical engineer. I created the constant thickness by relating the outer radius and the inner radius to the padding of the card with variable relations of 1.5*pad & 0.5*pad respectively.

I also had an issue where the padding of the child completion container and the padding of the parent opponent card were overlapping, making the "wall thickness" have a size of 0. This was fixed by switching the completion container's box-sizing property to border-box.